+++ src/index.ts
@@ -1,0 +1,7 @@
+import resolveProcessEnv from "./resolveProcessEnv";
+import retrieveSecret from "./retrieveSecret";
+
+export default {
+  retrieveSecret,
+  resolveProcessEnv
+}
