+++ src/lib/types.ts
@@ -1,0 +1,5 @@
+export interface SecretReference {
+  project: string;
+  name: string;
+  version: string;
+}
